chore(ci): satisfy lint gates for merge — 500-LoC splits, clippy, fmt, typos, rustdoc
- 500-LoC cap: split mtp_head.rs (634->406) by moving the DraftProposer impl to
  mtp_head/draft_proposer.rs; split sequence.rs (530->422) by moving save/restore
  state I/O to trait_impl/sequence/state_io.rs. Pure code motion — 216 spark-model
  unit tests unchanged, cargo check/doc clean.
- clippy: scope redundant_closure_call allow to the prefill profiling macro; use
  the unicode >= in a speculative.rs doc table to clear doc_lazy_continuation.
- typos: allowlist `nd` (the num-drafts notation nd=2/nd=3, not "and").
- rustdoc (surfaced by stricter rustdoc; main's green run predates it): qualify
  super::connect_kv_peer_backend, Self::take_drafter_kv, and delink the
  public->pub(crate) drafter_context reference.
- cargo fmt.
