+++ crates/spark-model/src/model/trait_impl/async_chkpt.rs
@@ -284,5 +284,4 @@ impl TransformerModel {
         self.gpu.record_event(self.secondary_event, stream)?;
         Ok(())
     }
-
 }
